improved Cheicken and fish AI

+++ Sources/Hevadea/Scenes/SceneGameplay.cs
@@ -47,8 +47,8 @@ namespace Hevadea.Scenes
         public override string GetDebugInfo()
         {
             return
-                $@"World time: {Game.World.DayNightCycle.Time}
-Time of the day: {Game.World.DayNightCycle.TimeOfTheDay} / {Game.World.DayNightCycle.CycleDuration}
+                $@"World time: {(int)Game.World.DayNightCycle.Time}
+Time of the day: {(int)Game.World.DayNightCycle.TimeOfTheDay} / {Game.World.DayNightCycle.CycleDuration}
 Days : {Game.World.DayNightCycle.DayCount}
 Current Stage: {Game.World.DayNightCycle.GetCurrentStage().Name} : {(int)Game.World.DayNightCycle.GetTimeOfTheCurrentStage()}/{(int)Game.World.DayNightCycle.GetCurrentStage().Duration}
 Player pos {Game.MainPlayer.X} {Game.MainPlayer.Y}
